using prompt installer

+++ install.py
@@ -1,34 +1,34 @@
 import os
-from config import CONFIG
+from config import *
 
 
 def configure_git():
-    resp1 = os.system("git config --global user.email \"%s\"" % CONFIG["email"])
-    resp2 = os.system("git config --global user.name \"%s\"" % CONFIG["name"])
+    print('----- Configuring GIT -----')
+    
+    confirm = input('Your are about to configure git. Do you want to continue ? Y/N : ')
+    if confirm.upper() in ('Y', 'YES') :
+        resp1 = os.system("git config --global user.email \"%s\"" % CONFIG["email"])
+        resp2 = os.system("git config --global user.name \"%s\"" % CONFIG["name"])
 
-    if resp1 + resp2 == 0:
-        print("git configuration .... OK")
-    else:
-        print("git configuration .... KO")
+        if resp1 + resp2 == 0:
+            print("git configuration .... OK")
+        else:
+            print("git configuration .... KO")
 
 
-def install_packages():
+def install_packages(packages):
     """
     Install useful packages
     :return:
     """
-    utils_packages = ("build-essential", "git", "tree", "git-flow", "vim", "maven", "supervisor", "php", "libapache2-mod-php", "apache2", "php-mcrypt",
-                      "mysql-server", "php-mysql", "phpmyadmin", "php-mbstring", "php-gettext", "php-sqlite3", "atool", "ipython", "python3-setuptools", "python3-mysqldb", "php-gettext", "php-xdebug", "libapache2-mod-php7.0",
-                      "ssh", "gimp", "curl", "terminator", "zsh", "npm", "vsftpd")
-
     status = []
-
     install_ok = 0  # number of installed packages
-
-    total = len(utils_packages)  # number of packages to install
+    total = len(packages)  # number of packages to install
 
     print("---- Starting packages installation  ----")
-    for package in utils_packages:
+    os.system("sudo apt-get update")
+
+    for package in packages:
         resp = os.system("sudo apt-get install %s -y" % package)
 
         # set installation status
@@ -55,19 +55,39 @@ def install_packages():
         print(s)
     print("Report: %d/%d packages installed" % (install_ok, total))
 
-
-# Installing JDK7 JDK8
-def install_oracle_jdk():
+def install_utils_packages():
     """
-    Install oracle version of JDK7 and JDK8
+    Install some utilities packages
     :return:
     """
-    os.system("sudo apt-get install -y software-properties-common")
-    os.system("sudo add-apt-repository ppa:webupd8team/java")
-    os.system("sudo apt-get update")
-    os.system("sudo apt-get install -y oracle-java7-installer oracle-java8-installer")
+    confirm = input('Your are about to install utils packages. Do you want to continue ? Y/N : ')
+    if confirm.upper() in ('Y', 'YES') :
+        install_packages(UTILS_PACKAGES)
 
-    print("Installing Java7 OK ...")
+def install_php():
+    """
+    Install some php packages
+    :return:
+    """
+    confirm = input('Your are about to install php. Do you want to continue ? Y/N : ')
+    if confirm.upper() in ('Y', 'YES') :
+        install_packages(PHP_PACKAGES)
+
+def install_java():
+    """
+    Install oracle version of JDK11 and JDK13
+    :return:
+    """
+    print('----- Installing Java -----')
+
+    confirm = input('Your are about to install java. Do you want to continue ? Y/N : ')
+    if confirm.upper() in ('Y', 'YES') :
+        os.system("sudo apt-get install -y software-properties-common")
+        os.system("add-apt-repository ppa:linuxuprising/java")
+        os.system("sudo apt-get update")
+        os.system("sudo apt-get install -y openjdk-8-jdk openjdk-11-jdk openjdk-13-jdk openjdk-14-jdk oracle-java11-installer oracle-java13-installer")
+
+        print("Installing java-11 and java13 OK ...")
 
 
 def install_npm_modules():
@@ -75,15 +95,21 @@ def install_npm_modules():
     Install some npm modules
     :return:
     """
-    os.system("sudo ln -sf /usr/bin/nodejs /usr/bin/node")
-    modules = ["npm", "yo", "gulp-cli", "bower", "generator-angular"]
+    print('----- Installing npm modules -----')
 
-    for mod in modules:
-        os.system("sudo npm install -g %s" % mod)
-    
-    os.system("sudo apt remove nodejs npm")
-    os.system("curl -sL https://deb.nodesource.com/setup_6.x | sudo -E bash -")
-    os.system("sudo apt-get install -y nodejs")
+    confirm = input('Your are about to install npm modules. Do you want to continue ? Y/N : ')
+    if confirm.upper() in ('Y', 'YES') :
+        os.system("sudo apt-get update")
+        os.system("sudo apt-get install -y npm")
+        os.system("sudo ln -sf /usr/bin/nodejs /usr/bin/node")
+        modules = ["npm", "yo", "gulp-cli", "bower", "generator-angular", "generator-jhipster"]
+
+        for mod in modules:
+            os.system("sudo npm install -g %s" % mod)
+
+        os.system("sudo apt remove nodejs npm")
+        os.system("curl -sL https://deb.nodesource.com/setup_6.x | sudo -E bash -")
+        os.system("sudo apt-get install -y nodejs")
 
 
 def configure_hosts():
@@ -91,9 +117,13 @@ def configure_hosts():
     add costum hosts
     :return:
     """
-    if 'dev.perform-world.com' not in open('/etc/hosts').read():
-        os.system("echo  '127.0.0.1       dev.perform-world.com' | sudo tee -a /etc/hosts")
-        os.system("echo  '127.0.0.1       local.supervisor' | sudo tee -a /etc/hosts")
+    print('----- Configuring hosts -----')
+
+    confirm = input('Your are about to configure hosts. Do you want to continue ? Y/N : ')
+    if confirm.upper() in ('Y', 'YES') :
+        if 'dev.perform-world.com' not in open('/etc/hosts').read():
+            os.system("echo  '127.0.0.1       dev.perform-world.com' | sudo tee -a /etc/hosts")
+            os.system("echo  '127.0.0.1       local.supervisor' | sudo tee -a /etc/hosts")
 
 
 def configure_apache2():
@@ -101,41 +131,99 @@ def configure_apache2():
     configures apache2
     :return:
     """
-    os.system("sudo cp ./apache2/sites-available/perform-world.conf /etc/apache2/sites-available")
-    os.system("sudo a2enmod proxy")
-    os.system("sudo service apache2 restart")
-    os.system("sudo a2ensite perform-world.conf")
+    print('----- Configuring apache2 -----')
+    confirm = input('Your are about to configure apache2. Do you want to continue ? Y/N : ')
+    if confirm.upper() in ('Y', 'YES') :
+        os.system("sudo cp ./apache2/sites-available/perform-world.conf /etc/apache2/sites-available")
+        os.system("sudo a2enmod proxy")
+        os.system("sudo service apache2 restart")
+        os.system("sudo a2ensite perform-world.conf")
+        os.system('sudo curl -L "https://github.com/docker/compose/releases/download/1.26.0/docker-compose-$(uname -s)-$(uname -m)" -o /usr/local/bin/docker-compose')
+        os.system("sudo chmod +x /usr/local/bin/docker-compose")
 
-
+def install_docker():
+    """
+    Install docker
+    """
+    print('----- Installing docker -----')
+    confirm = input('Your are about to install docker. Do you want to continue ? Y/N : ')
+    if confirm.upper() in ('Y', 'YES') :
+        os.system("sudo apt-get update")
+        os.system("sudo apt-get remove docker docker-engine docker.io")
+        os.system("sudo apt install -y docker.io")
+        os.system("sudo systemctl start docker")
+        os.system("sudo systemctl enable docker")
 
 def configure_supervisor():
+
     """
     configures supervisord
     :return:
     """
+    print('----- Installing supervisor -----')
 
-    os.system("sudo cp ./apache2/sites-available/supervisord.conf /etc/apache2/sites-available")
+    confirm = input('Your are about to install supervisor. Do you want to continue ? Y/N : ')
+    if confirm.upper() in ('Y', 'YES') :
+        os.system("sudo apt-get update")
+        os.system("sudo apt install -y supervisor")
+        os.system("sudo cp ./apache2/sites-available/supervisord.conf /etc/apache2/sites-available")
 
-    if 'kaba' not in open('/etc/hosts').read():
-        os.system("echo  '[inet_http_server]' | sudo tee -a /etc/supervisor/supervisord.conf")
-        os.system("echo  'port = 127.0.0.1:9001' | sudo tee -a /etc/supervisor/supervisord.conf")
-        os.system("echo  'username = kaba' | sudo tee -a /etc/supervisor/supervisord.conf")
-        os.system("echo  'password = {SHA}da3c01ea4729ba5fc5ed83d2e85b2b00c118753f' | sudo tee -a /etc/supervisor/supervisord.conf")
+        if 'kaba' not in open('/etc/hosts').read():
+            os.system("echo  '[inet_http_server]' | sudo tee -a /etc/supervisor/supervisord.conf")
+            os.system("echo  'port = 127.0.0.1:9001' | sudo tee -a /etc/supervisor/supervisord.conf")
+            os.system("echo  'username = kaba' | sudo tee -a /etc/supervisor/supervisord.conf")
+            os.system("echo  'password = {SHA}da3c01ea4729ba5fc5ed83d2e85b2b00c118753f' | sudo tee -a /etc/supervisor/supervisord.conf")
 
-    os.system("sudo a2ensite supervisord.conf")
-    os.system("sudo service apache2 restart")
-    os.system("sudo service supervisor restart")
+        os.system("sudo a2ensite supervisord.conf")
+        os.system("sudo service apache2 restart")
+        os.system("sudo service supervisor restart")
 
-def run():
-
-    os.system("sudo apt-get update")
-    install_packages()
+def install_all():
+    install_utils_packages()
     configure_git()
-    install_oracle_jdk()
-    install_npm_modules()
+    install_php()
     configure_hosts()
     configure_apache2()
-    # configure_supervisor()
+    configure_supervisor()
+    install_java()
+    install_npm_modules()
+    install_docker()
+
+def run():
+    menu = '''
+    --- Packages and tools intaller ---
+    1- Install utils packages
+    2- Configure GIT
+    3- Install PHP
+    4- Configure Hosts
+    5- Configure apache2
+    6- Configure supervisor
+    7- Install Java
+    8- Install NPM modules
+    9- Install Docker
+    10- Install all
+    11- Quit
+
+    Your choice :
+    '''
+    print(menu)
+    choice = int(input('Your choice : '))
+
+    choice_resolver = {
+        1: install_utils_packages,
+        2: configure_git,
+        3: install_php,
+        4: configure_hosts,
+        5: configure_apache2,
+        6: configure_supervisor,
+        7: install_java,
+        8: install_npm_modules,
+        9: install_docker,
+        10: install_all,
+        11: quit
+    }
+
+    choice_resolver[choice]()
 
 if __name__ == '__main__':
     run()
